chore: code cleanup
