+++ helpdesk/integrations/erpnext/utils.py
@@ -176,3 +176,17 @@ def _resync_data_fields(self_doctype: str, self_name: str, other_name: str) -> N
             hd_name,
             update_modified=False,
         )
+
+
+def find_target_for(doctype: str | None, value: str | None) -> tuple[str, str] | None:
+    """Map (HD Customer, name) ↔ (Customer, name) via the link fields on each
+    side. Returns None when no counterpart exists or inputs are missing."""
+    if not doctype or not value:
+        return None
+    if doctype == "HD Customer":
+        erp = frappe.db.get_value("Customer", {"hd_customer": value}, "name")
+        return ("Customer", erp) if erp else None
+    if doctype == "Customer":
+        hd = frappe.db.get_value("HD Customer", {"erpnext_customer": value}, "name")
+        return ("HD Customer", hd) if hd else None
+    return None
